state-search-be/pyproject.toml
[tool.poetry]
authors = ["timechess <tyc221@ruc.edu.cn>"]
description = ""
name = "state_search_be"
packages = [
    { include = "state_search_be" },
    { include = "state_search", from = "state_search_be" },
]
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
bleach = "^6.1.0"
fastapi-limiter = "^0.1.5"
grpcio = "^1.69.0"
grpcio-tools = "^1.69.0"
jsonlines = "^4.0.0"
prisma = "^0.15.0"
protobuf = "^5.29.3"
python = ">=3.11, <3.13"
python-dotenv = "^1.0.0"
sentence-transformers = "^3.3.1"

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]

[tool.ruff]
line-length = 88
target-version = "py311"

exclude = ["tests/", "**/__pycache__/", ".venv/", "dist/", "state_search"]